split route handling into separate instrumentation module
