+++ instrumentation/akka/akka-http-10.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/akkahttp/server/AkkaHttpServerInstrumentationModule.java
@@ -28,13 +28,14 @@ public class AkkaHttpServerInstrumentationModule extends InstrumentationModule {
   }
 
   @Override
+  public boolean isIndyModule() {
+    // AkkaHttpServerInstrumentationModule and AkkaHttpServerRouteInstrumentationModule share
+    // AkkaRouteHolder class
+    return false;
+  }
+
+  @Override
   public List<TypeInstrumentation> typeInstrumentations() {
-    return asList(
-        new HttpExtServerInstrumentation(),
-        new GraphInterpreterInstrumentation(),
-        new PathMatcherInstrumentation(),
-        new PathMatcherStaticInstrumentation(),
-        new RouteConcatenationInstrumentation(),
-        new PathConcatenationInstrumentation());
+    return asList(new HttpExtServerInstrumentation(), new GraphInterpreterInstrumentation());
   }
 }
