+++ instrumentation/akka/akka-http-10.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/akkahttp/server/route/RouteConcatenationInstrumentation.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.javaagent.instrumentation.akkahttp.server;
+package io.opentelemetry.javaagent.instrumentation.akkahttp.server.route;
 
 import static net.bytebuddy.matcher.ElementMatchers.namedOneOf;
